+++ src/highlight/mod.rs
@@ -5,7 +5,8 @@ use self::{
     c::highlight_c, cpp::highlight_cpp, css::highlight_css, elixir::highlight_elixir,
     go::highlight_go, html::highlight_html, java::highlight_java, javascript::highlight_javascript,
     json::highlight_json, lua::highlight_lua, php::highlight_php, python::highlight_python,
-    rust::highlight_rust, scala::highlight_scala,
+    rust::highlight_rust, scala::highlight_scala, tsx::highlight_tsx,
+    typescript::highlight_typescript,
 };
 
 mod bash;
@@ -24,6 +25,8 @@ mod php;
 mod python;
 mod rust;
 mod scala;
+mod tsx;
+mod typescript;
 
 static HIGHLIGHT_NAMES: [&str; 18] = [
     "attribute",
@@ -90,6 +93,8 @@ pub fn highlight_code(language: &str, lines: &[u8]) -> HighlightInfo {
         "python" => HighlightInfo::Highlighted(highlight_python(lines).unwrap()),
         "rust" => HighlightInfo::Highlighted(highlight_rust(lines).unwrap()),
         "scala" => HighlightInfo::Highlighted(highlight_scala(lines).unwrap()),
+        "tsx" => HighlightInfo::Highlighted(highlight_tsx(lines).unwrap()),
+        "typescript" | "ts" => HighlightInfo::Highlighted(highlight_typescript(lines).unwrap()),
         _ => HighlightInfo::Unhighlighted,
     }
 }